move react components to new 'components' directory
