+++ src/components/Crypto.js
