+++ src/components/SystemInfo.js
